pyproject.toml
[project]
name = "bytes"
version = "0.1.0"
description = "your friendly helper chatbot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.18",
    "lxml>=5.4.0",
    "openai>=1.81.0",
    "pandas>=2.2.3",
    "pdfminer-six>=20250506",
    "pgvector>=0.4.1",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pydantic-ai>=0.2.6",
    "pydantic[email]>=2.11.4",
    "pymupdf>=1.26.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.2",
    "langchain-chroma>=0.2.3",
    "langchain-huggingface>=0.2.0",
    "camelot-py[cv]>=1.0.0",
    "langchain-google-genai>=2.0.10",
    "langchain-core>=0.3.63",
    "passlib>=1.7.4",
    "pytest>=8.4.0",
    "langchain-experimental>=0.3.4",
    "plotly>=6.1.2",
]

[project.scripts]
bytes = "bytes.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
